Sql (#42)
* siquery: build dummy virtual table outside of rusqlite

* siquery: rename virtual table

* siquery: update to latest vtab branch

* siquery: update to latest vtab

* query a row table of vector of strings

* query multiple rows with specific headers

* query multiple tables in the same module

* register table in memory while executing a query && add: query tale header while creating schema

* register table in memory use the same data base to query tables

* run the first query before registering the tables && code cleaning

* loop into column names and print the query response. should fix : bad type error if querying a number

* delete todo line

* fix: make sur the table is not empty before getting a value by id

* code cleaning

* code cleaning

* merging with master

* WIP : save table in memory while creating the vtable

* create the table in memory when receiving the query && handling rusqlite errors

* fix : error handler of any user entry && code cleaning

* register schemas while initializing the db && fix querying values by type

* code cleaning && print result as pretty table

* code cleaning

* code cleaning

* update rusqlite vtab from ce9d325 rev (stable version)

* use rusqlite Value as type to fill table

* fix: typed tests

* fix: fields types for linux

*  fix: types fields for macOS

* WIP : fromat output to json && some code cleaning

* WIP : print tables into json & removed : function to query specific headers

* WIP : print tables into csv

* add : print into pretty & csv

* fix json

*  code cleaning

* WIP : add print mode flag to the sql query command

* add json/csv/pretty flags for table printer

* handle query errors

* fix error

* code cleaning

* WIP : command line flag to print all tables & execute static query with -A flag & code cleaning

* siquery table test & clean warnings

* code cleaning

* test tables methods

* code cleaning

* code cleaing

* fix macro nomenclature

+++ siquery_tables/src/linux/uptime.rs
@@ -13,7 +13,7 @@ impl Uptime {
             hours: 0,
             minutes: 0,
             seconds: 0,
-            total_seconds: 0.0,
+            total_seconds: 0,
         };
 
         let mut info: libc::sysinfo = unsafe { mem::zeroed() };
@@ -28,24 +28,24 @@ impl Uptime {
 
         if let Some(get_days) = Some(_t / sec_to_days_converter){
             remaining_time = _t - (get_days * sec_to_days_converter);
-            upt.days = get_days as u64;
+            upt.days = get_days as i64;
         }
 
         if let Some(get_hours) = Some(remaining_time / sec_to_hours_converter){
             remaining_time = remaining_time - (get_hours * sec_to_hours_converter);
-            upt.hours = get_hours as u64;
+            upt.hours = get_hours as i64;
         }
 
         if let Some(get_minutes) = Some(remaining_time / sec_to_minutes_converter){
             remaining_time = remaining_time - (get_minutes * sec_to_minutes_converter);
-            upt.minutes = get_minutes as u64;
+            upt.minutes = get_minutes as i64;
         }
 
         if let Some(get_seconds) = Some(remaining_time){
-            upt.seconds = get_seconds as u64;
+            upt.seconds = get_seconds as i64;
         }
 
-        upt.total_seconds = _t as f64;
+        upt.total_seconds = _t as i64;
 
         output.push(upt);
         output
